eliminate logback.py
