+++ myuw/logger/logresp.py
@@ -1,7 +1,6 @@
 import json
 import logging
-from myuw.logger.logback import log_info, log_time, log_exception_with_timer
-from myuw.logger.session_log import hash_session_key
+from myuw.logger.session_log import hash_session_key, get_userid
 
 resp_logger = logging.getLogger(__name__)
 
@@ -38,20 +37,38 @@ def log_err(logger, timer, exc_info):
     exc_info is a string containing
     the full stack trace, the exception type and value
     """
-    log_exception_with_timer(logger, timer, exc_info)
+    logger.error("{} {} Time={} seconds".format(get_userid(),
+                                                exc_info.splitlines(),
+                                                timer.get_elapsed()))
 
 
-def log_msg(logger, timer, msg):
-    log_time(logger, msg, timer)
+def log_exception(logger, action, exc_info):
+    """
+    exc_info is a string containing
+    the full stack trace, the exception type and value
+    """
+    logger.error("{} {} => {} ".format(get_userid(),
+                                       action,
+                                       exc_info.splitlines()))
+
+
+def log_info(logger, message):
+    logger.info("{} {}".format(get_userid(), message))
 
 
 def log_data_not_found_response(logger, timer):
-    log_time(logger, 'Data not found', timer)
+    log_msg(logger, timer, 'Data not found')
 
 
 def log_invalid_netid_response(logger, timer):
-    log_time(logger, 'Invalid netid, abort', timer)
+    log_msg(logger, timer, 'Invalid netid, abort')
 
 
 def log_invalid_regid_response(logger, timer):
-    log_time(logger, 'Invalid regid, abort', timer)
+    log_msg(logger, timer, 'Invalid regid, abort')
+
+
+def log_msg(logger, timer, action_message):
+    log_info(logger,
+             "{}. Time={} seconds".format(action_message,
+                                          timer.get_elapsed()))
